new/requirements.txt
feedparser
SQLAlchemy
celery
spacy
psycopg2-binary
